Conflicts:
	src/mess/drivers/camplynx.c

+++ src/emu/sound/multipcm.c
@@ -4,7 +4,7 @@
  * Sega System 32 Multi/Model 1/Model 2 custom PCM chip (315-5560) emulation.
  *
  * by Miguel Angel Horna (ElSemi) for Model 2 Emulator and MAME.
- * Information by R.Belmont and the YMF278B (OPL4) manual.
+ * Information by R. Belmont and the YMF278B (OPL4) manual.
  *
  * voice registers:
  * 0: Pan
@@ -158,7 +158,7 @@ void multipcm_device::EG_Calc(SLOT *slot)
 #define LFIX(v) ((unsigned int) ((float) (1<<LFO_SHIFT)*(v)))
 
 //Convert DB to multiply amplitude
-#define DB(v)   LFIX(pow(10.0,v/20.0))
+#define DB(v)   LFIX(powf(10.0f,v/20.0f))
 
 //Convert cents to step increment
 #define CENTS(v) LFIX(powf(2.0f,v/1200.0f))
